requirements.txt
django==1.11rc1
mysqlclient==1.3.10
python-memcached==1.58
python-decouple==3.0
pytz==2016.7
gunicorn==19.7.1
sinaweibopy==1.1.3
pillow==3.4.2
httplib2==0.9.2

dj-database-url==0.4.2
djangorestframework==3.6.2
django-qiniu-storage==2.3.0
django-tinymce==2.6.0
django-contrib-comments==1.8
django-autocomplete-light==3.2.1

